refactor: reorganize authentication routes and services
- Moved authentication routes into a dedicated package structure for better organization.
- Replaced the old auth routes with new modular routes for device authentication, password management, and email verification.
- Updated service dependencies to reflect the new structure, enhancing clarity and maintainability.
- Removed the legacy AuthService in favor of a more modular approach with specific services for credentials, device auth, password management, and email verification.
