+++ dependencies/__init__.py
@@ -34,16 +34,21 @@ from dependencies.infra import (
     get_settings,
 )
 from dependencies.services import (
+    fetch_user_profile,
     get_api_key_service,
     get_app_grant_repo,
-    get_auth_service,
     get_click_service,
     get_contact_service,
+    get_credential_service,
+    get_device_auth_service,
     get_export_service,
     get_oauth_service,
+    get_password_service,
     get_profile_picture_service,
     get_stats_service,
     get_url_service,
+    get_user_repo,
+    get_verification_service,
 )
 
 __all__ = [
@@ -60,23 +65,28 @@ __all__ = [
     "VerifiedUser",
     "check_api_key_scope",
     # services
+    "fetch_user_profile",
     "get_api_key_service",
     "get_app_grant_repo",
-    "get_auth_service",
     "get_click_service",
     "get_contact_service",
+    "get_credential_service",
     "get_current_user",
     # infra
     "get_db",
+    "get_device_auth_service",
     "get_email_provider",
     "get_export_service",
     "get_geoip_service",
     "get_oauth_service",
+    "get_password_service",
     "get_profile_picture_service",
     "get_redis",
     "get_settings",
     "get_stats_service",
     "get_url_service",
+    "get_user_repo",
+    "get_verification_service",
     "optional_scopes",
     "optional_scopes_verified",
     "require_auth",
