+++ services/auth/__init__.py
@@ -1,0 +1,33 @@
+"""
+Auth service package — split by domain concern.
+
+Re-exports all service classes and constants for convenient imports.
+"""
+
+from services.auth.credentials import CredentialService
+from services.auth.device import (
+    APP_ID_MAX_LEN,
+    DEVICE_AUTH_EXPIRY_SECONDS,
+    DeviceAuthService,
+)
+from services.auth.otp import (
+    MAX_TOKENS_PER_HOUR,
+    MAX_VERIFICATION_ATTEMPTS,
+    OTP_EXPIRY_SECONDS,
+    OtpService,
+)
+from services.auth.password import PasswordService
+from services.auth.verification import EmailVerificationService
+
+__all__ = [
+    "APP_ID_MAX_LEN",
+    "DEVICE_AUTH_EXPIRY_SECONDS",
+    "MAX_TOKENS_PER_HOUR",
+    "MAX_VERIFICATION_ATTEMPTS",
+    "OTP_EXPIRY_SECONDS",
+    "CredentialService",
+    "DeviceAuthService",
+    "EmailVerificationService",
+    "OtpService",
+    "PasswordService",
+]
